% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviator.R
\name{abbreviator}
\alias{abbreviator}
\title{Create Bootstrap 4 tooltip with an abbreviation and its definition}
\usage{
abbreviator(x, y)
}
\arguments{
\item{x}{the abbreviation, as string, matching the abbreviation data.table key}

\item{y}{the abbreviation data.table}
}
\value{
a html anchor (i.e. <a ...>KEY</a>) defining a tooltip with the abbreviation's definition (defined in data.table column "def").
}
\description{
Create Bootstrap 4 tooltip with an abbreviation and its definition
}
\examples{
\dontrun{
library(data.table)

abbrv <- data.table(abbrv = c("PGS"), def = c("Paul G Stevenson"), key = "abbrv")

abbreviator("PGS", abbrv)
}

}
