% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-biomass2.r
\docType{data}
\name{biomass2}
\alias{biomass2}
\title{Biomasa a nivel de árbol para cuatro especies arbóreas de Canadá}
\format{
Los datos contienen las siguientes columnas:
\describe{
\item{arbol}{Número del árbol.}
\item{spp}{Nombre común de la especie, como sigue: \verb{Balsam fir} es
\emph{Abies balsamea}, \verb{Black spruce} es \emph{Picea mariana},
\verb{White birch} es \emph{Betula papyrifera}, y \verb{White spruce}
es \emph{Picea glauca}.}
\item{dap}{Diámetro a la altura del pecho (1.3 m), en cm.}
\item{atot}{Altura total, en m.}
\item{wtot}{Biomasa total, en kg.}
\item{wfus}{Biomasa del fuste, en kg.}
\item{wramas}{Biomasa de las ramas, en kg.}
\item{whojas}{Biomasa del follaje, en kg.}
}
}
\source{
Los datos fueron cedidos cortesía del profesor Timothy Gregoire,
School of Forestry and Environmental Studies, Yale University (New
Haven, CT, USA).
}
\usage{
biomass2
}
\description{
Biomasa a nivel de árbol y otras variables, para cuatro
especies que crecen en bosques de Canadá.
}
\examples{
data(biomass2)
head(biomass2)
tapply(biomass2$wtot,biomass2$spp,summary)
}
\keyword{datasets}
