% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_REDCap.R
\name{checkbox_labels}
\alias{checkbox_labels}
\title{checkbox_labels}
\usage{
checkbox_labels(x, dict)

checkbox_labels(x, dict)
}
\arguments{
\item{x}{checkbox `Variable / Field Name` per data dictionary}

\item{dict}{REDCap data dictionary}
}
\value{
Named list of character objects

Named list of character objects
}
\description{
Apply factor labels to categorical responses (checkboxes) per REDCap data dictionary

Apply factor labels to categorical responses (checkboxes) per REDCap data dictionary
}
\examples{
\dontrun{

dat <- tibble(var___0 = c("0", "1"), var___1 = c("1", "0"))

dictionary <- tibble(`Variable / Field Name` = "var",
  `Choices, Calculations, OR Slider Labels` = "0, No | 1, Yes")

dat <- labelled::set_variable_labels(dat, .labels = checkbox_labels("var", dictionary))

}

\dontrun{

dat <- tibble(var___0 = c("0", "1"), var___1 = c("1", "0"))

dictionary <- tibble(`Variable / Field Name` = "var",
  `Choices, Calculations, OR Slider Labels` = "0, No | 1, Yes")

dat <- labelled::set_variable_labels(dat, .labels = checkbox_labels("var", dictionary))

}

}
