% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{copy_clip}
\alias{copy_clip}
\title{copy_clip}
\usage{
copy_clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{object to be copied}
}
\description{
Function to copy data from R to the clipboard.
}
\seealso{
\link[=paste_clip]{paste_clip()}
}
