% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_and_sort.R
\name{count_and_sort}
\alias{count_and_sort}
\title{Count and Sort Observations by Multiple Group Variables}
\usage{
count_and_sort(data, ...)
}
\arguments{
\item{data}{A data frame to be processed.}

\item{...}{Grouping variables passed to dplyr::group_by.}
}
\value{
A data frame with the counts of each group, sorted in descending order.
}
\description{
This function groups the data by the specified variables, counts the number of observations
in each group, and then sorts these groups in descending order based on the count.
}
\examples{
df <- data.frame(
  x = sample(LETTERS[1:5], 100, replace = TRUE),
  y = sample(1:3, 100, replace = TRUE)
)
count_and_sort(df, x, y)
}
