% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_functions.R
\name{create_REDCap_dictionary}
\alias{create_REDCap_dictionary}
\title{create_REDCap_dictionary}
\usage{
create_REDCap_dictionary(x, dictionary)
}
\arguments{
\item{x}{a dataframe with REDCap column names retained.}

\item{dictionary}{a REDCap dictionary (extracted from REDCap with standard formatting)}
}
\value{
object of class dataframe"
}
\description{
A function to generate a data dictionary for a REDCap dataset.
}
\examples{
questionnaire_dictionary <- exportMeta(redcap_connection)
create_REDCap_dictionary(questionnaire_year1, questionnaire_dictionary)


}
