% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cwd2.r
\docType{data}
\name{cwd2}
\alias{cwd2}
\title{Contiene mediciones de material leñoso muerto en una
unidad de muestreo de linea.}
\format{
Data contiene las siguientes columnas:
\describe{
\item{elemento}{Número del elemento (i.e., trozo) medido dentro
de la muesta de LIS.}
\item{diam}{Diámetro en la mitad del elemento, en cm.}
\item{lar}{Largo, en m.}
\item{ang}{Ángulo de intersección del elemento con la línea de
muestreo LIS, en grados.}
}
}
\source{
Datos digitados desde Marshall et al (2000).
}
\usage{
cwd2
}
\description{
Son variables medidas a nivel de trozo en la unidad de muestreo
(LIS).  El largo de la línea de muestreo o transecto es de
30 m.  Detalles sobre este tipo de estrategia de muestreo
se pueden revisar en  Gregoire and Valentine (2008).
}
\examples{
data(cwd2)
cwd2
}
\references{
\itemize{
\item Marshall PL, G Davis, VM LeMay. 2000. Using line intersect
sampling for coarse woody debris. Technical Report 3,
British Columbia Forest Service, Nanaimo, BC, Canada. 34 p.
\item Gregoire TG, HT Valentine. 2008. Sampling Strategies for Natural
Resources and the Environment. New York, USA.
Chapman & Hall/CRC. 474 p.
}
}
\keyword{datasets}
