% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excel_ref}
\alias{excel_ref}
\title{Excel column reference map}
\format{
A data table with 18278 rows and 2 columns:
\describe{
  \item{letters}{character Excel column reference letters}
  \item{numbers}{integer corresponding column number}
}
}
\usage{
data(excel_ref)
}
\description{
A data table of alpha/numeric key/value pairs, where key is the
letter sequence of the excel column. Range of letter sequence is A - ZZZ.
}
\keyword{datasets}
