% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expogralfull.r
\name{expogralfull.fx}
\alias{expogralfull.fx}
\title{Function that computes the result of the Stage  allometric
model.}
\usage{
expogralfull.fx(x, alpha, beta, gamma, delta, epsilon, upsilon = 0)
}
\arguments{
\item{x}{is the predictor variable.}

\item{alpha}{is the coefficient-parameter  \eqn{\alpha}.}

\item{beta}{is the  coefficient-parameter  \eqn{\beta}.}

\item{gamma}{is the  coefficient-parameter  \eqn{\gamma}.}

\item{delta}{is the  coefficient-parameter  \eqn{\delta}.}

\item{epsilon}{is the  coefficient-parameter  \eqn{\epsilon}.}

\item{upsilon}{is an optional constant term that force the prediction
of \emph{y} when \emph{x=0}. Thus, the new model becomes
\eqn{ y_i = \Upsilon+ f(x_i,\mathbf{\theta})}, where
\eqn{\mathbf{\theta}} is the vector of coefficients of
the above described function represented by
\eqn{f(\cdot)}. The default
value for \eqn{\Upsilon} is 0.}
}
\value{
Returns the response variable based upon
the predictor variable and the coefficients.
}
\description{
Function of the fully generalized-exponential allometric
model, based upon five  parameters (i.e., coefficients) and a variable,
as defined by the mathematical expression
\deqn{y_i= 
\alpha \mathrm{e}^{\left(\beta +\frac{\gamma}{x_i^{\delta}+\epsilon} \right)},
}
where: \eqn{y_i} and \eqn{x_i} are the response
and predictor variable, respectively, for the \emph{i}-th observation;
and the rest are parameters (i.e., coefficients).
Further details on this function can be found in
Salas-Eljatib (2025).
}
\examples{
# Predictor variable values to be used
time<-seq(5,60,by=0.01)
# Using the function
y<-expogralfull.fx(x=time,alpha=20,beta=.1,gamma=10,delta=1,
epsilon=1)
plot(time,y,type="l",col="red")
 
}
\references{
\itemize{
\item Salas-Eljatib C. 2025. Funciones alométricas: reparametrizaciones
y características matemáticas. Documento de trabajo No. 1,
Serie: Cuadernos de biometría, Laboratorio de Biometría y
Modelación Forestal, Universidad de Chile. Santiago, Chile. 51 p.
\url{https://biometriaforestal.uchile.cl}
}
}
\author{
Christian Salas-Eljatib.
}
