% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_REDCap.R
\name{factor_convert}
\alias{factor_convert}
\title{factor_convert}
\usage{
factor_convert(x, d, dict)

factor_convert(x, d, dict)
}
\arguments{
\item{x}{character vector}

\item{d}{REDCap import}

\item{dict}{REDCap data dictionary}
}
\description{
Apply factor labels to categorical responses per REDCap data dictionary.

Apply factor labels to categorical responses per REDCap data dictionary.
}
\examples{
\dontrun{

dat <- tibble(var = c("0", "1"))

dictionary <- tibble(`Variable / Field Name` = "var",
  `Choices, Calculations, OR Slider Labels` = "0, No | 1, Yes")

mutate(dat, across("var", factor_convert, d = dat, dict = dictionary))

}

\dontrun{

dat <- tibble(var = c("0", "1"))

dictionary <- tibble(`Variable / Field Name` = "var",
  `Choices, Calculations, OR Slider Labels` = "0, No | 1, Yes")

mutate(dat, across("var", factor_convert, d = dat, dict = dictionary))

}

}
