% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_REDCap.R
\name{factor_table}
\alias{factor_table}
\title{factor_table}
\usage{
factor_table(x, rows = "\\\\|", cols = ",")

factor_table(x, rows = "\\\\|", cols = ",")
}
\arguments{
\item{x}{character vector of REDCap options (e.g. 0, No | 1, Yes)}

\item{rows}{(default "\\|") Regex defining row separator}

\item{cols}{(default ",") Regex defining column separator}
}
\value{
2 column tibble of factor levels ("key") and labels ("value")

2 column tibble of factor levels ("key") and labels ("value")
}
\description{
factor_table

factor_table
}
\examples{
\dontrun{ factor_table("0, No | 1, Yes") }
\dontrun{ factor_table("0, No | 1, Yes") }

}
