% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_model.R
\name{format_model}
\alias{format_model}
\title{Format a model for quick in-text referencing}
\usage{
format_model(x, effects = NULL, conf.level = 0.95, dp = 2)
}
\arguments{
\item{x}{a model object}

\item{effects}{(default NULL) a tidy effects option e.g. "fixed", "random"}

\item{conf.level}{(default 0.95) the reported confidence interval of model estimates}

\item{dp}{(default 2) the number of decimal points reported in returned strings}
}
\value{
a tibble with default `tidy()` parameters plus `conf.low`, `conf.high`, `text1`, and `text2`
}
\description{
This function builds on `broom::tidy()` by adding formatted strings ready for use in text.
}
\examples{
\dontrun{
data(iris)
m1 <- lm(Sepal.Length ~ Sepal.Width, data = iris)

format_model(m1) \%>\% select(term, estimate, conf.low, conf.high)

format_model(m1)$text1
format_model(m1)$text2
}

}
