% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_text.R
\name{format_text}
\alias{format_text}
\title{format_text}
\usage{
format_text(x, conf.level = 0.95, dp = 2)
}
\arguments{
\item{x}{a tibble with at least columns named estimate, conf.low and conf.high}

\item{conf.level}{the label describing the calculated confidence level}

\item{dp}{the number of decimal places to report}
}
\value{
text1: estimate (95\% CI: lower to upper)

text2: (estimate; 95\% CI: lower to upper)

text3 (for tables): estimate <br /> CI [lower, upper]
}
\description{
Converts model output from quick_text() into a tibble with 3 pre-formatted options.
}
\seealso{
\link[=quick_text]{quick_text()}
}
