% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.rr.R
\name{glm.rr}
\alias{glm.rr}
\title{glm.rr}
\usage{
glm.rr(mod, digits = 2)
}
\arguments{
\item{mod}{output from a glm model (with family binomial or Poisson)}

\item{digits}{(default 2) the number of decimal points to be used across the output}
}
\value{
Model output, in summary form
}
\description{
Returns summary table output for a glm model, taking care of exponentiating the coefficients and confidence intervals
}
\examples{
\dontrun{
 counts <- c(18,17,15,20,10,20,25,13,12)
 outcome <- gl(3,1,9)
 treatment <- gl(3,3)
 data.frame(treatment, outcome, counts) # showing data
 glm.D93 <- glm(counts ~ outcome + treatment, family = poisson())

 summary(glm.D93)
 glm.rr(glm.D93)
}

}
