% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtree.r
\name{gtree}
\alias{gtree}
\title{Computes basal area of any given tree}
\usage{
gtree(x, in.m2)
}
\arguments{
\item{x}{is the vector having tree diameter, in cm}

\item{in.m2}{is an indicator variable: 1 to obtain the result in m2 if x was measured
in cm; and 0 to obtain the resulting area in the same units of x.}
}
\value{
The value of basal area.
}
\description{
Computes basal area of any given tree. Actually provides the area for a given circle of radius x.
}
\details{
No details are given
}
\examples{

#Creating an example dataframe
dbh<- round(rnorm(4,25,20),1);
#Using the function
in.m2=1
gtree(dbh,in.m2)

}
\author{
Christian Salas-Eljatib
}
