% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A function to generate a data dictionary for a REDCap dataset.}
\usage{
hello(text)
}
\arguments{
\item{text}{some text}
}
\value{
text
}
\description{
A function to generate a data dictionary for a REDCap dataset.
}
\examples{
\dontrun{
questionnaire_dictionary <- exportMeta(redcap_connection)
create_REDCap_dictionary(questionnaire_year1, questionnaire_dictionary)
}

}
