% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.a.r
\name{interp.a}
\alias{interp.a}
\title{Linear interpolation for three data of x and y only works if the first y is missing.}
\usage{
interp.a(xs, ys)
}
\arguments{
\item{xs}{Vector number of size 3}

\item{ys}{Vector number of size 3, with first position empty or NA}
}
\value{
Output description (simple).
}
\description{
Linear interpolation
}
\details{
Linear interpolation for three data of x and y only works if the first y is missing.
}
\examples{
x<-c(1,2,3)
y<-c(NA,4,6)
interp.a(x,y)
}
\author{
Christian Salas.
}
