% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.b.r
\name{interp.b}
\alias{interp.b}
\title{Linear interpolation for three data of x and y only works if the second y is missing.}
\usage{
interp.b(xs, ys)
}
\arguments{
\item{xs}{Vector number of size 3}

\item{ys}{Vector number of size 3, with second position empty or NA}
}
\value{
Output description (simple).
}
\description{
Linear interpolation
}
\details{
Linear interpolation for three data of x and y only works if the second y is missing.
}
\examples{
x<-c(1,2,3)
y<-c(4,NA,6)
interp.b(x,y)
}
\author{
Christian Salas.
}
