% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\name{meansd}
\alias{meansd}
\title{meansd}
\usage{
meansd(x, dpp = 2, dps = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{dpp}{(default 2) the number of decimal points to return the mean}

\item{dps}{(default 2) the number of decimal points to return the standard deviation}
}
\value{
A string the form of "mean (sd)" at the specified precision
}
\description{
Return the mean and sd of a vector
}
\examples{
\dontrun{
  data(mtcars)

  meansd(mtcars$cyl)
}

}
