% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianiqr.R
\name{medianiqr}
\alias{medianiqr}
\title{medianiqr}
\usage{
medianiqr(x, dpp = 2, dps = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{dpp}{(default 2) the number of decimal points to return for the median}

\item{dps}{(default 2) the number of decimal points to return for the interquartile range}
}
\value{
object of class character, in the form "XX.XX (XX.XX, XX.XX)"
}
\description{
Return the median and interquartile range of a vector
}
\examples{
  medianiqr(mtcars$mpg)

}
