% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nele.list.r
\name{nele.list}
\alias{nele.list}
\title{Extract the n-th element from a list}
\usage{
nele.list(lst, n)
}
\arguments{
\item{lst}{is the list object}

\item{n}{is the position of the element in the list to be retrieved}
}
\value{
object with elements of each list
}
\description{
Extract the n-th element from a list
}
\examples{
x <- list(list("z","x","y"), list(3,4,99,23,45), list(1,67,5,6,89))
nele.list(x,1)
nele.list(x,2)
nele.list(x,3)

}
\author{
Christian Salas-Eljatib
}
