% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPer.R
\name{nPer}
\alias{nPer}
\title{nPer}
\usage{
nPer(vec, intGroup, dp = 1)
}
\arguments{
\item{vec}{numeric vector}

\item{intGroup}{xxxx}

\item{dp}{(default 1)}
}
\value{
xxxx
}
\description{
Printing N and percentage
}
\examples{
\dontrun{
  nPer(xxx)
}

}
