% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{paste_clip}
\alias{paste_clip}
\title{paste_clip}
\usage{
paste_clip(header = TRUE, ...)
}
\arguments{
\item{header}{default TRUE}

\item{...}{other arguements passed to read.table}
}
\description{
Function to paste data from clipboard into R.
}
\seealso{
\link[=copy_clip]{copy_clip()}
}
