% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gbs.R
\name{print_gbs}
\alias{print_gbs}
\title{print_gbs}
\usage{
print_gbs(x, kable = F)
}
\arguments{
\item{x}{the output, via piping, of a summarise statement that follows a group_by (tidyverse)}

\item{kable}{(default F) add kable and kable_styling to the output (for using within markdown documents)}
}
\value{
A table (either `tibble` or `kable` based)
}
\description{
Print the output from a gb (group_by) s (summarise). The purpose of this function is to aid in making a common table, with groups as columns and summary statistics in the cells. When creating this via `group_by() %>% summarise()` you need to transpose the output and correct the column headings, this function does that.
}
\examples{
\dontrun{
mtcars \%>\%
  group_by(cyl) \%>\%
  summarise(cyl_n = n(),
            mean_hp = mean(hp)) \%>\%
  print_gbs()
}

}
