% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd.r
\name{qmd}
\alias{qmd}
\title{Computes the quadratic mean diameter of a sample plot.}
\usage{
qmd(tph = tph, gha = gha)
}
\arguments{
\item{tph}{is tree density, in trees/ha;}

\item{gha}{is basal area, in m2/ha}
}
\value{
Returns the quadratic mean diameter (in cm) for a given plot.
}
\description{
This function computes the quadratic mean diameter of a sample plot.
}
\examples{

#using the function
qmd(tph=1023, gha=50)
}
\author{
Christian Salas-Eljatib.
}
