% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_text.R
\name{quick_text}
\alias{quick_text}
\title{quick_text}
\usage{
quick_text(
  mod_list,
  effects = NULL,
  conf.int = TRUE,
  exponentiate = FALSE,
  conf.level = 0.95,
  dp = 2
)
}
\arguments{
\item{mod_list}{a list of regression models}

\item{effects}{a broom::tidy option e.g. "fixed", "random"}

\item{conf.int}{a broom::tidy option to generate confidence intervals, logical}

\item{exponentiate}{a broom::tidy optoin, logical}

\item{conf.level}{the reported confidence interval of model estimates}

\item{dp}{the number of decimal points reported in returned strings}
}
\value{
a list of tibbles generated from broom:tidy.

text1: estimate (95\% CI: lower to upper)

text2: (estimate; 95\% CI: lower to upper)

text3 (for tables): estimate <br /> CI [lower, upper]
}
\description{
Improvements to biometrics::format_model.
}
\seealso{
\link[=format_text]{format_text()}
}
