% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder.R
\name{recoder}
\alias{recoder}
\title{Recode a Text File as UTF-8}
\usage{
recoder(x)
}
\arguments{
\item{x}{path to the file}
}
\description{
Recode a Text File as UTF-8
}
\examples{

\dontrun{

recoder("packages.bib")

}

}
