% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_up.R
\name{roll_up}
\alias{roll_up}
\title{Combine all variable values and separated for each key group in a data frame - not sorted.}
\usage{
roll_up(x, vars = NULL)
}
\arguments{
\item{x}{...}

\item{vars}{an optional character vector of the desired output column names, this uses the \code{variables} function to create the vector. If omitted, variables names will be sequentially indexed variables of name \emph{V}.}
}
\value{
A data frame is returned with a id variable (matching the first column of the input data frame) and columns for values.
}
\description{
\code{roll_up} for grouped members of a data frame, column 1, combine all the values that are listed in all variables and separated into new columns.

For example, combine all diagnosis values for a patient that has multiple entries in a dataset and return a new diagnosis list with each value in a defined column.
}
\examples{
\dontrun{
temp %>% group_by(root) %>%
    do(roll_up(., vars = c("diagnosis", "dagger", "ediag")))
}

}
