% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
<<<<<<< HEAD
\title{Rounds the double columns in a data frame}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a data frame object}

\item{digits}{the number of digits after the decimal point}
}
\value{
the data frame with rounded numbers
}
\description{
All columns in a data frame with class double will be rounded to the sepecified
number of digits. Integer classes will be left unchanged.
}
\examples{

df <- data.frame(a = "a", b = 0.025, c = 15.34234, d = T)

round_df(df, 2)
=======
\title{Round data frame decimal points}
\usage{
round_df(df, dp = 2)
}
\arguments{
\item{df}{the data frame to be rounded}

\item{dp}{(default 2) the desired number of decimal points}
}
\value{
a data frame, the rounded columns are convertd to character objects.
}
\description{
Round the numeric columns of a data frame to the specified number of decimal
points using the `sprintf` function.
}
\examples{
\dontrun{
data(mtcars)

head( round_df(mtcars, 1) )
}
>>>>>>> 5f675318c758311140ae387150324351c1da0776

}
