% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_vec}
\alias{round_vec}
\title{round_vec}
\usage{
round_vec(x, digits = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\value{
object of class character"
}
\description{
A function, for vectors, to keep trailing zeroes when rounding numbers.
}
\examples{
round_vec(mtcars$wt, 2)


}
