% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundf.R
\name{roundf}
\alias{roundf}
\title{roundf}
\usage{
roundf(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\value{
object of class character"
}
\description{
A function to keep trailing zeroes when rounding numbers.
}
