% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_telethonkids.R
\name{scale_color_telethonkids}
\alias{scale_color_telethonkids}
\title{Color scale constructor for Telethon Kids colours.}
\usage{
scale_color_telethonkids(palette = "light", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in telethonkids_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for Telethon Kids colours.
}
