% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_thekids.R
\name{scale_color_thekids}
\alias{scale_color_thekids}
\title{Color scale constructor for The Kids colours.}
\usage{
scale_color_thekids(palette = "light", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in thekids_palettes. Options are "primary", "tint50" and "typography"}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for The Kids colours.
}
