% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_telethonkids.R
\name{scale_fill_telethonkids}
\alias{scale_fill_telethonkids}
\title{Fill scale constructor for Telethon Kids colors.}
\usage{
scale_fill_telethonkids(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in telethonkids_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for Telethon Kids colors.
}
