% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_thekids.R
\name{scale_fill_thekids}
\alias{scale_fill_thekids}
\title{Fill scale constructor for The Kids colors.}
\usage{
scale_fill_thekids(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tehkids_palettes. Options are "primary", "tint50" and "typography"}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for The Kids colors.
}
