% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standtab.r
\name{standtab}
\alias{standtab}
\title{Function to creates a stand for a given imputed-tree-list of
a sample plot.}
\usage{
standtab(
  data,
  plot.area,
  y.for.class,
  yclass.amp,
  metric = TRUE,
  want.basal.area = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame having the tree list of a sample plot.}

\item{plot.area}{column name having the plot area in m\eqn{^{2}}{^2}.}

\item{y.for.class}{a text of the column-name having the tree-size variable to
be used to segregated the tree density.}

\item{yclass.amp}{a numeric value establishing the width of each
class of the variable \code{y.for.class}. The measurement unit must be
the same as the one of the variable \code{y.for.class}.}

\item{metric}{is a logic value, the default is to \code{TRUE}, thus
the diameter \code{d} has to be in cm, and the resulting tree-level
basal area will be in m\eqn{^{2}}{^2}. If \code{metric} is \code{FALSE},
the diameter \code{d} has to be in inches, and the computed tree
basal area will be in ft\eqn{^{2}}{^2}.}

\item{want.basal.area}{is a logic value, if set to \code{TRUE},
the stand basal area, segregated by the classes of the\code{y.for.class}
is added as a column, next to the tree density one. The basal area
column by class is expressed in m\eqn{^{2}}{^2}/ha. The default
for this option is \code{FALSE}.}

\item{...}{aditional options for basic stats functions.}
}
\value{
Returns a stand table.
}
\description{
Creates a stand table, i.e. a table that describes the tree density
of a forest segregated by a tree-size variable \code{y}. The most common
example is to use diameter at breast-heigth (i.e., \eqn{d}) as the
variable, thus the resulting table is a way to describe the
diameter distribution.
}
\details{
For a given imputed tree list of a sample plot, tree
density is computed by classes of the tree-size variable. Note that
the imputed-tree list must have all the tree-level variables
needed to compute stand-level variables, such as, diameter and plot
size.
}
\examples{

## Dataframe to be used
df<-biometrics::eucaplot2
#see the metadata by typing ?eucaplot2
head(df)
datana::descstat(df[,c("dap","atot")])
## Preparing the treelist, in the required format
df$parce<-1;df$sup.plot<-500;amp.diam=2;
out<-standtab(data=df,plot.area="sup.plot",y.for.class = "dap",
              yclass.amp = amp.diam)
out
#the following sum
sum(out$nha.class)
# must be equal to the stand-level density (in trees/ha)
df$expf<-10000/df$sup.plot;sum(df$expf)
# Do you want to add the stand basal area segregated by `y.for.class`
out2<-standtab(data=df,plot.area="sup.plot",y.for.class = "dap",
               yclass.amp = amp.diam, want.basal.area = TRUE)
out2
#the following sum
sum(out2$gha.class)
# must be equal to the stand-level basal area (in m2/ha)
df$g<-gtree(x=df$dap);df$gtree.ha<-df$g*df$expf;sum(df$gtree.ha)
}
\references{
\itemize{
\item Salas-Eljatib C. 2025. Biometría y Modelación Forestal.
Borrador de libro, en revisión. 352 p.
}
}
\author{
Christian Salas-Eljatib.
}
