% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telethonkids_colours}
\alias{telethonkids_colours}
\title{Telethon Kids Institute colours}
\format{
A named vector key = pantone matching system colour, value = hex
    colour code
}
\usage{
data(telethonkids_colours)
}
\description{
Hex codes for the Telethon Kids colours per the 2014 Brand Style Guide.
}
\keyword{datasets}
