% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telethonkids_cols.R
\name{telethonkids_cols}
\alias{telethonkids_cols}
\title{Function to extract Telethon Kids colors as hex codes}
\usage{
telethonkids_cols(...)
}
\arguments{
\item{...}{Character names of telethonkids_colours}
}
\description{
Function to extract Telethon Kids colors as hex codes
}
