% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telethonkids_pal.R
\name{telethonkids_pal}
\alias{telethonkids_pal}
\title{Return function to interpolate a Telethon Kids color palette}
\usage{
telethonkids_pal(palette = "light", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in telethonkids_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Telethon Kids color palette
}
