% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telethonkids_palettes}
\alias{telethonkids_palettes}
\title{Telethon Kids Institute Palettes}
\format{
\describe{
  \item{light}{Light Colour Tone Palette}
  \item{dark}{Dark Colour Tone Palette}
  \item{grey}{Typography Palette}
}
}
\usage{
data(telethonkids_palettes)
}
\description{
Telethon Kids colour palettes per the 2014 Brand Style Guide.
}
\keyword{datasets}
