% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thekids_colours}
\alias{thekids_colours}
\title{The Kids colours}
\format{
A named vector key = pantone matching system colour, value = hex
    colour code
}
\usage{
data(thekids_colours)
}
\description{
Hex codes for the The Kids colours per the 2024 Brand Style Guide.
}
\keyword{datasets}
