% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thekids_cols.R
\name{thekids_cols}
\alias{thekids_cols}
\title{Function to extract The Kids colors as hex codes}
\usage{
thekids_cols(...)
}
\arguments{
\item{...}{Character names of thekids_colours}
}
\description{
Function to extract The Kids colors as hex codes
}
