% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thekids_pal.R
\name{thekids_pal}
\alias{thekids_pal}
\title{Return function to interpolate a The Kids color palette}
\usage{
thekids_pal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in thekids_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a The Kids color palette
}
