% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thekids_palettes}
\alias{thekids_palettes}
\title{The Kids Palettes}
\format{
\describe{
  \item{primary}{Primary Colours Palette}
  \item{tint50}{50 Tint Palette}
  \item{grey}{Typography Palette}
}
}
\usage{
data(thekids_palettes)
}
\description{
The Kids colour palettes per the 2024 Brand Style Guide.
}
\keyword{datasets}
