% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thekids_table.R
\name{thekids_table}
\alias{thekids_table}
\title{The Kids themed table output}
\usage{
thekids_table(
  x,
  font.size = 8,
  font.size.header = 10,
  line.spacing = 1.5,
  padding = 2,
  colour = "CoolGrey"
)
}
\arguments{
\item{x}{a table, typically a data.frame, tibble, or output from gtsummary}

\item{font.size}{the font size for text in the body of the table, defaults to 8 (passed throught to set_flextable_defaults)}

\item{font.size.header}{the font size for text in the header of the table, defaults to 10}

\item{line.spacing}{line spacing for the table, defaults to 1.5 (passed throught to set_flextable_defaults)}

\item{padding}{padding around all four sides of the text within the cell, defaults to 2 (passed throught to set_flextable_defaults)}

\item{colour}{a colour platte from The Kids branding, options include "Saffron", "Pumpkin", "Teal", "DarkTeal", "CelestialBlue", "AzurBlue", "MidnightBlue", or "CoolGrey", defaults to 'CoolGrey'}
}
\value{
a flextable class object that will display in both html and word output
}
\description{
Using flextable (as this originated with a focus on word documents), this applies The Kids branding as a theme to table output, include the use of the Barlow font.
}
\details{
It also has default settings which produce a relatively compact table, to avoid reports becoming excessively lengthy.
}
\examples{

\dontrun{

head(mtcars, 10) \%>\%
  thekids_table(colour = "Saffron")
}

}
