% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_institute.R
\name{theme_institute}
\alias{theme_institute}
\title{theme_institute
Applies themes and colour schemes for universal ggplot2 formatting within the
biostatistics team. Barlow Semi Condensed Font Family can be downloaded
through Google Fonts}
\usage{
theme_institute()
}
\description{
theme_institute
Applies themes and colour schemes for universal ggplot2 formatting within the
biostatistics team. Barlow Semi Condensed Font Family can be downloaded
through Google Fonts
}
