% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-treevolruca2.r
\docType{data}
\name{treevolruca2}
\alias{treevolruca2}
\title{Volumen a nivel de árbol para especies nativas del bosque de Rucamanque}
\format{
Las siguientes columnas son parte de la dataframe:
\describe{
\item{arbol}{Número del árbol.}
\item{spp}{Codificación de la especie como sigue:
"Laurel" es \emph{Laurelia sempervirens},
"Lingue" es \emph{Persea lingue},
"Olivillo" es \emph{Aextoxicon punctatum},
"Tepa" es \emph{Laureliopsis philippiana},
"Tineo" es \emph{Weinmannia trichosperma}, y
"Ulmo" es \emph{Eucryphia cordifolia}.}
\item{dap}{Diámetro a la altura del pecho, en cm.}
\item{atot}{Altura total, en m.}
\item{d6}{Diámetro fustal a los 6 m, en cm.}
\item{vtot}{Volumen bruto total, en m³ con corteza.}
}
}
\source{
Los datos fueron cedidos por el Dr. Christian Salas-Eljatib,
Universidad de Chile (Santago, Chile).
}
\usage{
data(treevolruca2)
}
\description{
Volumen, altura y diámetro, entre otras para árboles muestra
en el bosque de Rucamanque, cerca de Temuco,
en la región de la Araucanía, en el sur de Chile.
}
\examples{
data(treevolruca2)
head(treevolruca2)
table(treevolruca2$spp)
}
\references{
\itemize{
\item Salas C. 2002. Ajuste y validación de ecuaciones de volumen
para un relicto del bosque de Roble-Laurel-Lingue.
Bosque 23(2): 81-92. \doi{10.4067/S0717-92002002000200009}
\url{https://eljatib.com/publication/2002-07-01_ajuste_y_validacion_/}
}
}
\keyword{datasets}
