% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_REDCap.R
\name{variable_labels}
\alias{variable_labels}
\title{variable_labels}
\usage{
variable_labels(d, dict)

variable_labels(d, dict)
}
\arguments{
\item{d}{REDCap import}

\item{dict}{REDCap data dictionary}
}
\value{
character vector (length `ncol(d)`) of REDCap Field Labels per the data dictionary

data frame with variable labels
}
\description{
Apply variable labels to a data frame per the REDCap data dictionary. If
there is no variable label associated with a column then is lable will be the
same as the column's name.

Apply variable labels to a data frame per the REDCap data dictionary. If
there is no variable label associated with a column then is label will be the
same as the column's name.
}
\examples{
\dontrun{

dat <- tibble(var = 1)

dictionary <- tibble(`Variable / Field Name` = "var", `Field Label` = "Label")

dat <- labelled::set_variable_labels(dat, .labels = variable_labels(dat, dictionary))

}

\dontrun{

dat <- tibble(var = 1, var2___1 = 1, var2___2 = 0)
dictionary <- tibble(`Variable / Field Name` = c("var", "var2"),
                     `Field Label` = c("Label for field 'Var'", NA),
                     `Field Type` = c(NA, "checkbox"),
                     `Choices, Calculations, OR Slider Labels` =
                       c(NA, "1, First checkbox label | 2, Second checkbox label"))

dat <- variable_labels(dat, dictionary)
View(dat)

}

}
