% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Create a string vector with a numbered repeating final element.}
\usage{
variables(names, n)
}
\arguments{
\item{names}{string vector of any length, the last entry will be repeated \code{n} times with sequential labelling starting at 1.}

\item{n}{integer indicating the number of times that the last entry of \code{names} shall be repeated.}
}
\description{
\code{variables} returns a string vector where the last entry is repeated by a specified number of times with appended incremental numbering.
}
\examples{
variables(c("a", "b", "c"), 20)

}
