% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightheight_zscore.R
\name{weightheight_zscore}
\alias{weightheight_zscore}
\title{weightheight_zscore}
\usage{
weightheight_zscore(age, sex, length, weight, percentile = FALSE)
}
\arguments{
\item{age}{the age of the child (days)}

\item{sex}{the sex of the child ("M" = male, "F" = female)}

\item{length}{the length of the child (cm)}

\item{weight}{the weight of the child (kg)}

\item{percentile}{(default FALSE) return the raw Z-score for the given weight and length}
}
\value{
object of class double
}
\description{
Calculate the weight-for-length percentile for children aged 0--2 or 2--5.
}
\details{
Function created from the standard weight-for-length tables released by the WHO (https://www.who.int/tools/child-growth-standards/standards/weight-for-length-height).
The calculation of the Z-score is done using the methodology of http://www.ncbi.nlm.nih.gov/pubmed/32012066.

Limitations exist on the height of the child for the respective age classifications. Children aged 0--2 take the admissible height values between 45cm and 110cm (boys and girls). Children aged 2--5 take the admissible height values between 65cm and 120cm (boys and girls).

The function rounds the height value to the nearest 0.5cm, to save interpolation.
}
\examples{
weightheight_zscore(376, "M", 50, 8, percentile = FALSE)

}
