% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_table.R
\name{contingency_table}
\alias{contingency_table}
\title{contingency_table}
\usage{
contingency_table(vec1, vec2, dp = 1)
}
\arguments{
\item{vec1}{a vector - should be discrete values}

\item{vec2}{a vector - should be discrete values}

\item{dp}{number of decimal places in proportion}
}
\value{
object of class matrix
}
\description{
returns a frequency table of two vectors with both counts and proportions.
}
\examples{
#'
\dontrun{ contingency_table(mtcars$gear, mtcars$cyl) }#'
}
