% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{excel_ref}
\alias{excel_ref}
\title{Excel column reference map}
\format{A data table with 18278 rows and 2 columns:
\describe{
  \item{\code{letters}}{character Excel column reference letters}
  \item{\code{numbers}}{integer corresponding column number}
}}
\usage{
data(excel_ref)
}
\description{
A data table of alpha/numeric key/value pairs, where key is the
letter sequence of the excel column. Range of letter sequence is A - ZZZ.
}
\examples{
\donttest{
  excel_ref["ABC"]$numbers
}
}
\keyword{datasets}
