% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_interval.R
\name{group_interval}
\alias{group_interval}
\title{Group by date interval}
\usage{
group_interval(start, end, buffer = 0, .view_mat = F, .view_graph = F)
}
\arguments{
\item{start, end}{vectors of the start and end dates/times}

\item{buffer}{the buffer interval e.g. days(14), defaults to 0}

\item{.view_mat}{when TRUE (defult FALSE) creates an overlap matrix and applies
View()}

\item{.view_graph}{when TRUE (default FALSE) plots the graph of crouped records}
}
\value{
an integer vector of group numbers
}
\description{
Group records in a data frame by date that intervals including a buffer period.
}
\examples{

\dontrun{

dat <- tibble(
  name = c("Paul", "Jess", "Ben", "Josie", "Damien", "Hayley"),
  start = ymd(c("2018-05-25", "2018-05-29", NA, "2018-06-25",
    "2019-07-15", "2019-07-08")),
  end = ymd(c("2018-05-30", "2018-06-03", "2018-06-23",
    "2018-06-29", "2019-07-18", "2019-07-13")))

group_interval(dat$start, dat$end, days(2))

mutate(dat, group = group_interval(start, end, days(2)))

}

}
