% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_helper.R
\name{html_report}
\alias{html_report}
\title{html_report}
\usage{
html_report(toc = TRUE)
}
\arguments{
\item{toc}{default to TRUE to show menu bar}
}
\description{
Remove pandoc configuration from .Rmd.
}
