% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICD10}
\alias{ICD10}
\title{ICD-10 hash table}
\format{
A data frame with 6820 rows and 5 variables:
\describe{
  \item{id}{character key ICD-10 codes WITHOUT decimal points}
  \item{name}{character disease name}
  \item{disease_category}{character disease category}
  \item{parent_category}{character parent disease category}
  \item{disease_group}{character Infections/notifiable}
  \item{record_type}{ICD-10 or ICD-10-AM}
}
}
\source{
\url{http://health.gnu.org/}
}
\usage{
data(ICD10)
}
\description{
A data.table that maps ICD-10 classification code to the disease
name, disease category, parent category and disease group. Codes were
sourced from GNU Health (downloaded June-2018).
}
\keyword{datasets}
