% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPer.R
\name{nPer}
\alias{nPer}
\title{nPer}
\usage{
nPer(vec, intGroup, dp = 1)
}
\arguments{
\item{vec}{numeric vector}

\item{intGroup}{the response to be searched for (integer, character string, or vector of responses)}

\item{dp}{(default 1)}
}
\value{
object of class character, in the form "N (XX.X%)"
}
\description{
A helper function for use in descriptive statistics tables.
}
\details{
Prints N and percentage for categorical variables (accepts vector of responses)
}
