% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_df}
\alias{round_df}
\title{round_df}
\usage{
round_df(x, digits = 2, con_char = F)
}
\arguments{
\item{x}{a data frame or tibble vector}

\item{digits}{integer indicating the number of decimal places to be used}

\item{con_char}{logical as to whether you want the rounded columns to be converted to character class (T) or not (F); default is F}
}
\value{
object of class tibble"
}
\description{
A function, for data frames or tibbles, to round all numeric columns in a data frame or tibble.
}
\details{
The option exists to keep these columns as numeric (simply rounding), or to convert them to characters keeping trailing zeroes in the process.
}
\examples{

\dontrun{
round_df(iris, 0)

round_df(iris, 0, con_char = T) \%>\% str
}

}
