% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Rounds the double columns in a data frame}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a data frame object}

\item{digits}{the number of digits after the decimal point}
}
\value{
the data frame with rounded numbers
}
\description{
All columns in a data frame with class double will be rounded to the sepecified
number of digits. Integer classes will be left unchanged.
}
\examples{

df <- data.frame(a = "a", b = 0.025, c = 15.34234, d = T)

round_df(df, 2)

}
