% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Round data frame decimal points}
\usage{
round_df(df, dp = 2)
}
\arguments{
\item{df}{the data frame to be rounded}

\item{dp}{(default 2) the desired number of decimal points}
}
\value{
a data frame, the rounded columns are convertd to character objects.
}
\description{
Round the numeric columns of a data frame to the specified number of decimal
points using the `sprintf` function.
}
\examples{
\dontrun{
data(mtcars)

head( round_df(mtcars, 1) )
}

}
