% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telethonkids_colours.R
\docType{data}
\name{telethonkids_colours}
\alias{telethonkids_colours}
\title{Telethon Kids Institute colours per the 2014 Brand Style Guide.}
\format{An object of class \code{character} of length 22.}
\usage{
telethonkids_colours
}
\description{
Telethon Kids Institute colours per the 2014 Brand Style Guide.
}
\keyword{datasets}
