% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telethonkids_palettes.R
\docType{data}
\name{telethonkids_palettes}
\alias{telethonkids_palettes}
\title{Telethon Kids Colour palettes per the 2014 Brand Style Guide.}
\format{An object of class \code{list} of length 3.}
\usage{
telethonkids_palettes
}
\description{
Telethon Kids Colour palettes per the 2014 Brand Style Guide.
}
\keyword{datasets}
