% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thekids_table.R
\name{thekids_table}
\alias{thekids_table}
\title{Themed table output}
\usage{
thekids_table(
  x,
  font.size = 8,
  font.size.header = 10,
  line.spacing = 1.5,
  padding = 2,
  colour = "CoolGrey"
)
}
\arguments{
\item{start, end}{vectors of the start and end dates/times}

\item{buffer}{the buffer interval e.g. days(14), defaults to 0}

\item{.view_mat}{when TRUE (defult FALSE) creates an overlap matrix and applies
View()}

\item{.view_graph}{when TRUE (default FALSE) plots the graph of crouped records}
}
\value{
an integer vector of group numbers
}
\description{
Using flextable (as this originated with a focus on word documents), this applies theming to table output in line with The Kids branding
}
\examples{

\dontrun{

head(mtcars, 10) \%>\%
  thekids_table(colour = "Saffron")
}

}
